pyproject.toml
[project]
name = "exe-pgr-rounders"
version = "0.1.0"
description = "A full-stack web app to manage a summer rounders tournament"
requires-python = ">=3.11"
dependencies = [
    "blinker==1.7.0",
    "click==8.1.7",
    "flask==3.0.3",
    "flask-login==0.6.3",
    "flask-sqlalchemy==3.1.1",
    "greenlet==3.0.3",
    "itsdangerous==2.2.0",
    "jinja2==3.1.3",
    "markupsafe==2.1.5",
    "numpy==1.26.4",
    "pandas==2.2.2",
    "pillow>=11.2.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2024.1",
    "six==1.16.0",
    "sqlalchemy==2.0.29",
    "typing-extensions==4.11.0",
    "tzdata==2024.1",
    "werkzeug==3.0.2",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["rounders"]
namespaces = false

[tool.basedpyright]
typeCheckingMode = "basic"

